+++ circuits/spawn/spawn.circom
@@ -3,4 +3,4 @@ pragma circom 2.1.1;
 include "spawn.func.circom";
 
 component main { public [ canSpawn, spawnCityId, commitBlockHash, hPrevTile, 
-    hSpawnTile, hSecret ] } = Spawn();
+    hSpawnTile, hBlind ] } = Spawn();
