+++ circuits/spawn/spawn.smoke.json
@@ -4,7 +4,7 @@
     "commitBlockHash": "467454691729242243759967775001677318986529874328564320584076052117170924846",
     "hPrevTile": "19692098879346355348825224807794470761368065655776046974338573026213703765842",
     "hSpawnTile": "14613183192131249614094278234551057114893448565188121930735592481452777813187",
-    "hSecret": "19780519184160579587292741653074321551360276629440664611523018255743639027652",
+    "hBlind": "19780519184160579587292741653074321551360276629440664611523018255743639027652",
     "prevTile": [
         "0",
         "1",
@@ -23,5 +23,5 @@
         "0",
         "2"
     ],
-    "secret": "6683317219253039888235400875520003218434422633446391293015473481786261950392"
+    "blind": "6683317219253039888235400875520003218434422633446391293015473481786261950392"
 }
