+++ contracts/src/SpawnVerifier.sol
@@ -43,26 +43,26 @@ contract Groth16Verifier {
     uint256 constant deltay2 = 8495653923123431417604973247489272438418190587263600148770280649306958101930;
 
     
-    uint256 constant IC0x = 3739434496224682390786900921223997134940204635749720114746105168432237314157;
-    uint256 constant IC0y = 315664074963060319231567619575564183000281711750734892889062251987981671957;
+    uint256 constant IC0x = 14337957236479593781523959191758935376326338216437575446185713421900167401758;
+    uint256 constant IC0y = 20774398907388929156993407607910466830100128745326419042746014280026477065376;
     
-    uint256 constant IC1x = 470667422033536251624958427432932786715608424093607043761474515803777454351;
-    uint256 constant IC1y = 11281084176178103242340459022946232955457694493478858797733170052259557658408;
+    uint256 constant IC1x = 234547712551454137340518667887528437001244953233045463728900341929046815016;
+    uint256 constant IC1y = 20926951241347756964460696789346349128767585659446426703811719449761854985842;
     
-    uint256 constant IC2x = 21820817406899303620827604983207775316886104962411016826982800998395186281485;
-    uint256 constant IC2y = 7528304584200431015748973239460403516409670023593501625789879727446787355213;
+    uint256 constant IC2x = 18588288274791558385875777391985986402159171621475547255735075765790160056315;
+    uint256 constant IC2y = 596708158959891334500433585263722824551005064551526828815547413991111100155;
     
-    uint256 constant IC3x = 12912285310141144772477666570904266100376239087905005637815901909324165856919;
-    uint256 constant IC3y = 3450323293513543702168066615548109914052818897593487706315608264171833431551;
+    uint256 constant IC3x = 12243670187238028837071132144977479502292326200402049567492593932531574173432;
+    uint256 constant IC3y = 1368224953819552666766028079056963768212666703484372830272302555746951181089;
     
-    uint256 constant IC4x = 17984166045441818029000068892914432927974368048417739235897541397290570212515;
-    uint256 constant IC4y = 21674016687091230606316517744966344266205988711167265004800703455341778203364;
+    uint256 constant IC4x = 833603857391787296156210974110658259923473692512526313177071505017284433454;
+    uint256 constant IC4y = 9724984713831274923884593285588055176460762052107174847694682890545476545742;
     
-    uint256 constant IC5x = 7729802899703498498301378220663899623629388771331076674342766633625915889915;
-    uint256 constant IC5y = 16153469266678518487223358693456319582580116427578236368325797731997983150755;
+    uint256 constant IC5x = 18894469278435617246247957933925261886935829128124119244832049062940858896460;
+    uint256 constant IC5y = 21851864242664016649964638833800072063382589804539895258658715051715899188581;
     
-    uint256 constant IC6x = 377607984765380298302097591840847172568182200796604233795235714672247123587;
-    uint256 constant IC6y = 20223147914404463185227401275141433415794205785887946037338714216134358920636;
+    uint256 constant IC6x = 15829449195686874906028859296578397173668783640268833168064831053155063582017;
+    uint256 constant IC6y = 9196375556862508000292159963950546521964121144994214614376830186121005131868;
     
  
     // Memory data
