+++ contracts/src/codegen/world/ISpawnSystem.sol
@@ -20,7 +20,7 @@ interface ISpawnSystem {
 
   function getSpawnCommitment(address player) external view returns (uint256);
 
-  function getSpawnChallengeHash(address player) external view returns (uint256);
+  function getSpawnblindHash(address player) external view returns (uint256);
 
   function getBlockHash(uint256 blockCommited) external view returns (uint256);
 }
