+++ contracts/src/common/SpawnInputs.sol
@@ -7,5 +7,5 @@ struct SpawnInputs {
     uint256 commitBlockHash;
     uint256 hPrevTile;
     uint256 hSpawnTile;
-    uint256 hSecret;
+    uint256 hBlind;
 }
